pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "clawkit"
dynamic = ["version"]
description = "Free local social media content extractor"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "JoeyZ1105" }]
dependencies = ["httpx>=0.27"]

[project.optional-dependencies]
auth = ["playwright>=1.40"]
video = ["yt-dlp>=2024.0.0"]
analyze = ["google-genai"]

[project.scripts]
clawkit = "clawkit.cli:main"

[tool.setuptools]
include-package-data = true

[tool.setuptools.dynamic]
version = { attr = "clawkit.__version__" }

[tool.setuptools.packages.find]
include = ["clawkit*"]
